@@ -6,35 +6,35 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  flutter_staggered_animations: ^0.1.2
-  local_auth: ^0.6.3+4
-  shared_preferences: ^0.5.12+4
-  intl: ^0.16.1
-  flutter_svg:
-  platform_date_picker: ^0.0.4+1
-  pull_to_refresh: 1.6.3
-  dio: ^3.0.10
-  json_serializable: ^3.5.1
-  mobx: ^1.2.1+4
-  auto_route: ^0.6.9
-  injectable: ^1.0.7
-  get_it: ^5.0.6
-  flutter_mobx: ^1.1.0+2
-  oktoast: ^2.3.2
+  flutter_staggered_animations: ^1.0.0
+  local_auth: ^1.1.5
+  shared_preferences: ^2.0.5
+  intl: ^0.17.0
+  flutter_svg: ^0.22.0
+  platform_date_picker: ^0.2.0-nullsafety.0+1
+  pull_to_refresh: ^2.0.0
+  dio: ^4.0.0
+  json_serializable: ^4.1.1
+  mobx: ^2.0.1
+  flutter_mobx: ^2.0.0
+  auto_route: ^2.2.0
+  injectable: ^1.4.0
+  get_it: ^7.1.3
+  oktoast: ^3.0.0
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
-  auto_route_generator: ^0.6.10
-  injectable_generator: ^1.0.7
-  mobx_codegen: ^1.1.2
-  build_runner: ^1.11.0
-  faker: ^1.3.0
-  mockito: ^4.1.4
+  auto_route_generator: ^2.1.0
+  injectable_generator: ^1.4.0
+  mobx_codegen: ^2.0.1+3
+  build_runner: ^2.0.2
+  faker: ^2.0.0-rc.2
+  mockito: ^5.0.7
   flutter_test:
     sdk: flutter
 
