@@ -1,4 +1,4 @@
-import 'package:estudo_app/src/domain/models/result/post.dart';
+import 'package:estudo_app/src/domain/entities/post.dart';
 import 'package:estudo_app/src/ui/utils/helpers/date_time_helper.dart';
 import 'package:estudo_app/src/ui/utils/helpers/retry_helper.dart';
 import 'package:estudo_app/src/ui/utils/helpers/toast_helper.dart';
@@ -14,9 +14,12 @@ abstract class _PostControllerBase with Store {
   final DateTimeHelper dateTimeHelper;
   final RetryHelper retryHelper;
   final ToastHelper toastHelper;
-  Post post;
-  DateTime dateTime;
-  String text;
+  Post? post;
+  @observable
+  DateTime? dateTime;
+
+  @observable
+  String? text;
 
   _PostControllerBase(
     this.dateTimeHelper,
@@ -28,6 +31,16 @@ abstract class _PostControllerBase with Store {
     this.post = post;
   }
 
+  @action
+  void setText(String value) {
+    text = value;
+  }
+
+  @action
+  void setDateTime(DateTime value) {
+    dateTime = value;
+  }
+
   Future<void> doRetry() async {
     toastHelper.show('Oi');
     await retryHelper.testRetry(timeToRetry: 2, maximumAttempts: 4);
@@ -35,15 +48,14 @@ abstract class _PostControllerBase with Store {
   }
 
   void showText() {
-    toastHelper.show('$text-${dateTimeHelper.formatToDMY(dateTime)}');
+    toastHelper.show('$text-${dateTimeHelper.formatToDMY(dateTime!)}');
   }
 
   @computed
   bool get canShowValue {
+    // if (dateTime.toString().isNotEmpty && (text != null || text!.isNotEmpty)) {
     if (dateTime != null && text != null) {
-      if (dateTime.toString().isNotEmpty && text.isNotEmpty) {
-        return true;
-      }
+      return true;
     }
     return false;
   }
