@@ -1,9 +1,9 @@
-import 'package:auto_route/auto_route.dart';
 import 'package:estudo_app/config/router/base_router.gr.dart';
 import 'package:estudo_app/src/domain/use_cases/get_fingerprint_is_active/get_fingerprint_is_active_use_case_interface.dart';
+import 'package:estudo_app/src/ui/utils/helpers/fingerprint_helper.dart';
+import 'package:estudo_app/src/ui/utils/helpers/navigator_helper.dart';
 import 'package:estudo_app/src/ui/utils/helpers/toast_helper.dart';
 import 'package:injectable/injectable.dart';
-import 'package:local_auth/local_auth.dart';
 import 'package:mobx/mobx.dart';
 
 part 'splash_controller.g.dart';
@@ -14,62 +14,35 @@ class SplashController = _SplashControllerBase with _$SplashController;
 abstract class _SplashControllerBase with Store {
   final ToastHelper toastHelper;
   final IGetFingerprintIsActiveUseCase getFingerprintIsActiveUseCase;
-  final message = 'Autentique para entrar no app.';
+  final NavigatorHelper _navigatorHelper;
+  final FingerprintHelper _fingerprintHelper;
 
   @observable
-  String errorMessage;
+  String? errorMessage;
 
   _SplashControllerBase(
     this.toastHelper,
     this.getFingerprintIsActiveUseCase,
+    this._navigatorHelper,
+    this._fingerprintHelper,
   );
 
   @action
   void setErrorMessage(String errorMessage) => this.errorMessage = errorMessage;
 
   Future<void> checkBiometric() async {
-    final fingerprintIsActive =  getFingerprintIsActiveUseCase();
+    final fingerprintIsActive = getFingerprintIsActiveUseCase();
     if (fingerprintIsActive) {
-      final _localAuth = LocalAuthentication();
-      final canCheckBiometrics = await _localAuth.canCheckBiometrics;
-      if (canCheckBiometrics) {
-        final availableBiometrics = await _localAuth.getAvailableBiometrics();
-        if (availableBiometrics.contains(BiometricType.face)) {
-          try {
-            final didAuthenticate = await _localAuth.authenticateWithBiometrics(
-              localizedReason: message,
-            );
-            if (didAuthenticate) {
-              ExtendedNavigator.root.pushAndRemoveUntil(Routes.postsPage, (route) => false);
-            } else {
-              setErrorMessage('Não Autenticou.');
-              return;
-            }
-          } on Exception catch (_) {
-            setErrorMessage('Não Autenticou.');
-          }
-        }
-        if (availableBiometrics.contains(BiometricType.fingerprint)) {
-          try {
-            final didAuthenticate = await _localAuth.authenticateWithBiometrics(
-              localizedReason: message,
-            );
-            if (didAuthenticate) {
-              ExtendedNavigator.root.pushAndRemoveUntil(Routes.postsPage, (route) => false);
-            } else {
-              setErrorMessage('Não Autenticou.');
-              return;
-            }
-          } on Exception catch (_) {
-            setErrorMessage('Não Autenticou.');
-          }
-        }
-      } else {
-        setErrorMessage('Não existe biometria para usar.');
-        ExtendedNavigator.root.pushAndRemoveUntil(Routes.postsPage, (route) => false);
+      final fingerprintGranted = await _fingerprintHelper.checkBiometric();
+      if (fingerprintGranted) {
+        _navigatorHelper.router.pushAndPopUntil(PostsRoute(), predicate: (route) => false);
+        return;
       }
+      setErrorMessage('Não Autenticou.');
     } else {
-      ExtendedNavigator.root.pushAndRemoveUntil(Routes.postsPage, (route) => false);
+      Future.delayed(Duration(milliseconds: 200), () {
+        _navigatorHelper.router.pushAndPopUntil(PostsRoute(), predicate: (route) => false);
+      });
     }
   }
 }
