@@ -1,56 +1,58 @@
 import 'package:estudo_app/src/ui/utils/constants/app_colors.dart';
-import 'package:estudo_app/src/ui/utils/constants/app_text_styles.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/svg.dart';
 
 class AppTextInput extends StatefulWidget {
-  final Function onChanged;
-  final Function onSaved;
+  final Function(String)? onChanged;
   final String labelText;
-  final FocusNode currentFocusNode;
-  final FocusNode nextFocusNode;
-  final TextInputAction textInputAction;
-  final Function onFieldSubmitted;
-  final TextInputType textInputType;
-  final bool obscureText;
-  final Function validator;
-  final String initialValue;
+  final Function(String?)? onSaved;
+  final FocusNode? currentFocusNode;
+  final FocusNode? nextFocusNode;
+  final TextInputAction? textInputAction;
+  final Function? onFieldSubmitted;
+  final TextInputType? textInputType;
+  final bool? obscureText;
+
+  // final Function(String) validator;
+  final String? initialValue;
+  final String? errorText;
+  final String? value;
   final bool readOnly;
-  final String value;
-  final TextStyle textStyle;
-  final String suffixIconPath;
-  final Function suffixIconPathOnTap;
-  final InputBorder border;
-  final bool autoFocus;
-  final TextAlign textAlign;
-  final Function onTap;
-  final TextEditingController controller;
-  final List<TextInputFormatter> textInputFormatters;
+  final TextStyle? textStyle;
+  final String? suffixIconPath;
+  final Function()? suffixIconPathOnTap;
+  final InputBorder? border;
+  final bool? autoFocus;
+  final TextAlign? textAlign;
+  final Function()? onTap;
+  final TextEditingController? controller;
+  final List<TextInputFormatter>? textInputFormatters;
 
   const AppTextInput({
+    required this.labelText,
+    this.onChanged,
+    this.errorText,
     this.textInputAction,
     this.textInputType,
     this.value,
     this.initialValue,
     this.controller,
-    this.onChanged,
-    this.labelText,
     this.textStyle,
     this.currentFocusNode,
     this.nextFocusNode,
-    this.obscureText = false,
-    this.validator,
+    // this.validator,
     this.onFieldSubmitted,
-    this.readOnly = false,
     this.onSaved,
     this.border,
     this.textInputFormatters,
     this.suffixIconPath,
-    this.textAlign = TextAlign.start,
-    this.autoFocus = false,
     this.suffixIconPathOnTap,
     this.onTap,
+    this.obscureText = false,
+    this.readOnly = false,
+    this.textAlign = TextAlign.start,
+    this.autoFocus = false,
   }) : super();
 
   @override
@@ -58,25 +60,48 @@ class AppTextInput extends StatefulWidget {
 }
 
 class _AppTextInputState extends State<AppTextInput> {
-  bool showEye;
+  bool? showEye;
+
+  @override
+  Widget build(BuildContext context) => TextFormField(
+        controller: widget.controller,
+        decoration: InputDecoration(
+          border: widget.border,
+          errorText: widget.errorText,
+          alignLabelWithHint: true,
+          focusedBorder: UnderlineInputBorder(
+            borderSide: BorderSide(color: AppColors.primaryColor),
+          ),
+          enabledBorder: UnderlineInputBorder(
+            borderSide: BorderSide(color: AppColors.primaryColor),
+          ),
+          hintText: '',
+          labelText: widget.labelText,
+          suffixIcon: widget.obscureText! ? _setEye() : _setIcon(),
+        ),
+        onTap: widget.onTap,
+        readOnly: widget.readOnly,
+        onChanged: widget.onChanged,
+      );
 
   @override
   void initState() {
     super.initState();
-    showEye = widget.obscureText;
+    showEye = widget.obscureText!;
   }
 
-  @override
+  /* @override
   Widget build(BuildContext context) => Container(
         child: TextFormField(
-          textAlign: widget.textAlign,
+          textAlign: widget.textAlign!,
           controller: widget.controller,
-          autofocus: widget.autoFocus,
+          autofocus: widget.autoFocus!,
           onTap: widget.onTap,
           decoration: InputDecoration(
             border: widget.border,
-            errorText:
-                widget.validator == null || !widget.currentFocusNode.hasFocus ? null : widget.validator(widget.value),
+            // errorText: widget.validator == null || !widget.currentFocusNode!.hasFocus
+            //     ? null
+            //     : widget.validator(widget.value!),
             alignLabelWithHint: true,
             focusedBorder: UnderlineInputBorder(
               borderSide: BorderSide(color: AppColors.primaryColor),
@@ -86,33 +111,33 @@ class _AppTextInputState extends State<AppTextInput> {
             ),
             hintText: '',
             labelText: widget.labelText,
-            suffixIcon: widget.obscureText ? _setEye() : _setIcon(),
+            suffixIcon: widget.obscureText! ? _setEye() : _setIcon(),
           ),
           onChanged: widget.onChanged,
-          obscureText: showEye,
-          readOnly: widget.readOnly,
-          validator: widget.validator,
-          style: _setTextStyle(),
+          obscureText: showEye!,
+          readOnly: widget.readOnly!,
+          // validator: widget.validator,
+          style: _setTextStyle()!,
           textInputAction: widget.textInputAction,
           focusNode: widget.currentFocusNode,
           keyboardType: widget.textInputType,
-          onSaved: widget.onSaved,
+          onSaved: widget.onSaved!,
           inputFormatters: widget.textInputFormatters,
           onFieldSubmitted: (term) async {
             if (widget.onFieldSubmitted != null) {
-              await widget.onFieldSubmitted.call();
+              await widget.onFieldSubmitted!.call();
             }
-            _fieldFocusChange(context, widget.currentFocusNode, widget.nextFocusNode);
+            _fieldFocusChange(context, widget.currentFocusNode!, widget.nextFocusNode!);
           },
         ),
-      );
+      );*/
 
-  Widget _setEye() {
-    if (widget.obscureText) {
+  Widget? _setEye() {
+    if (widget.obscureText!) {
       return GestureDetector(
         onTap: () {
           setState(() {
-            showEye = !showEye;
+            showEye = !showEye!;
           });
         },
         child: Container(
@@ -120,7 +145,7 @@ class _AppTextInputState extends State<AppTextInput> {
           height: 10,
           width: 10,
           padding: const EdgeInsets.only(left: 15, right: 15),
-          child: showEye
+          child: showEye!
               ? Icon(
                   Icons.remove_red_eye_sharp,
                   color: AppColors.black,
@@ -135,7 +160,7 @@ class _AppTextInputState extends State<AppTextInput> {
     return null;
   }
 
-  Widget _setIcon() {
+  Widget? _setIcon() {
     if (widget.suffixIconPathOnTap != null && widget.suffixIconPath != null) {
       return InkWell(
         onTap: widget.suffixIconPathOnTap,
@@ -144,7 +169,7 @@ class _AppTextInputState extends State<AppTextInput> {
           width: 10,
           padding: const EdgeInsets.only(left: 15, right: 15),
           child: SvgPicture.asset(
-            widget.suffixIconPath,
+            widget.suffixIconPath!,
             alignment: Alignment.bottomCenter,
           ),
         ),
@@ -153,21 +178,21 @@ class _AppTextInputState extends State<AppTextInput> {
     return null;
   }
 
-  void _fieldFocusChange(BuildContext context, FocusNode currentFocus, FocusNode nextFocus) {
-    currentFocus.unfocus();
-
-    if (nextFocus != null) {
-      FocusScope.of(context).requestFocus(nextFocus);
-    }
-  }
-
-  TextStyle _setTextStyle() {
-    if (widget.textStyle != null) {
-      return widget.textStyle;
-    }
-    if (widget.readOnly) {
-      return AppTextStyle.textDisable;
-    }
-    return AppTextStyle.textBlack;
-  }
+// void _fieldFocusChange(BuildContext context, FocusNode currentFocus, FocusNode nextFocus) {
+//   currentFocus.unfocus();
+//
+//   if (null != nextFocus) {
+//     FocusScope.of(context).requestFocus(nextFocus);
+//   }
+// }
+//
+// TextStyle? _setTextStyle() {
+//   if (widget.textStyle != null) {
+//     return widget.textStyle;
+//   }
+//   if (widget.readOnly!) {
+//     return AppTextStyle.textDisable;
+//   }
+//   return AppTextStyle.textBlack;
+// }
 }
