@@ -1,3 +1,4 @@
+import 'package:estudo_app/src/ui/utils/helpers/navigator_helper.dart';
 import 'package:injectable/injectable.dart';
 import 'package:mobx/mobx.dart';
 
@@ -6,4 +7,8 @@ part 'app_controller.g.dart';
 @Injectable()
 class AppController = _AppControllerBase with _$AppController;
 
-abstract class _AppControllerBase with Store {}
+abstract class _AppControllerBase with Store {
+  final NavigatorHelper navigatorHelper;
+
+  _AppControllerBase(this.navigatorHelper);
+}
