@@ -1,3 +1,4 @@
+import 'package:estudo_app/config/router/base_router.gr.dart';
 import 'package:injectable/injectable.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -5,6 +6,5 @@ import 'package:shared_preferences/shared_preferences.dart';
 abstract class RegisterModule {
   @preResolve
   Future<SharedPreferences> get prefs => SharedPreferences.getInstance();
-
-  // Dio get dio => Dio();
+  final AppRouter appRouter = AppRouter();
 }
