@@ -1,5 +1,5 @@
 import 'package:estudo_app/src/domain/data_sources/get_posts/get_posts_datasource_interface.dart';
-import 'package:estudo_app/src/domain/models/result/post.dart';
+import 'package:estudo_app/src/domain/models/post_model.dart';
 import 'package:estudo_app/src/domain/utils/app_http_client/app_http_client_interface.dart';
 import 'package:injectable/injectable.dart';
 
@@ -10,8 +10,8 @@ class GetPostsDataSource implements IGetPostsDataSource {
   final IAppHttpClient _appHttpClient;
 
   @override
-  Future<List<Post>> call() async {
+  Future<List<PostModel>> call() async {
     final response = await _appHttpClient.get('/posts');
-    return (response.data as List).map((item) => Post.fromJson(item)).toList();
+    return (response.data as List).map((item) => PostModel.fromJson(item)).toList();
   }
 }
