@@ -1,4 +1,4 @@
-import 'package:estudo_app/src/domain/models/result/post.dart';
+import 'package:estudo_app/src/domain/entities/post.dart';
 
 abstract class IGetPostsRepository {
   Future<List<Post>> call();
