@@ -4,7 +4,11 @@ import 'package:injectable/injectable.dart';
 
 @injectable
 class LoggerHelper {
-  void show(String text) {
-    log('[LOGGER HELPER] $text');
+  void show(String text, {String? title}) {
+    if (title != null) {
+      log('[$title] $text');
+      return;
+    }
+    log('[APP_LOG] $text');
   }
 }
