@@ -1,6 +1,7 @@
 import 'dart:math' as math;
 
-import 'package:estudo_app/src/domain/models/result/post.dart';
+import 'package:estudo_app/src/domain/entities/post.dart';
+import 'package:estudo_app/src/domain/models/post_model.dart';
 import 'package:faker/faker.dart';
 
 mixin MockedObjects {
@@ -9,6 +10,13 @@ mixin MockedObjects {
 
   static Post getMockedPost() {
     return Post(
+      id: _faker.randomGenerator.integer(100),
+      title: _faker.randomGenerator.string(10),
+    );
+  }
+
+  static PostModel getMockedPostModel() {
+    return PostModel(
       userId: _faker.randomGenerator.integer(100),
       id: _faker.randomGenerator.integer(100),
       title: _faker.randomGenerator.string(10),
